+++ presentation/src/main/java/com/sms/presentation/main/ui/mypage/component/project/ProjectComponent.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.sms.presentation.main.ui.detail.data.ExpandableProjectData
+import com.sms.presentation.main.ui.mypage.state.ExpandableProjectData
 import com.sms.presentation.main.ui.detail.data.RelatedLinksData
 
 @Composable
